plugins/{alsa,pulse}: read() and write() return ssize_t

+++ plugins/alsa/alsa.c
@@ -99,7 +99,7 @@ l_wake_up(lua_State *L)
     LuastatusPluginData *pd = lua_touserdata(L, lua_upvalueindex(1));
     Priv *p = pd->priv;
 
-    int unused = write(p->self_pipe[1], "", 1); // write '\0'
+    ssize_t unused = write(p->self_pipe[1], "", 1); // write '\0'
     (void) unused;
 
     return 0;
@@ -309,7 +309,7 @@ iteration(LuastatusPluginData *pd, LuastatusPluginRunFuncs funcs)
             goto error;
         }
         if (pollfds.nprefix && (pollfds.data[0].revents & POLLIN)) {
-            int unused = read(p->self_pipe[0], (char [1]) {'\0'}, 1);
+            ssize_t unused = read(p->self_pipe[0], (char [1]) {'\0'}, 1);
             (void) unused;
         }
